+++ source/cpu/8080.cpp
@@ -21,14 +21,16 @@ SOFTWARE.
 */
 
 #include <assert.h>
+#include <format>
 #ifdef ENABLE_NLOHMANN_JSON
 #include <nlohmann/json.hpp>
 #else
 #define ARDUINOJSON_ENABLE_STRING_VIEW 1
 #include <ArduinoJson.h>
 #endif // ENABLE_NLOHMANN_JSON
-#include "meen/utils/Utils.h"
+
 #include "meen/cpu/8080.h"
+#include "meen/utils/Utils.h"
 #include "meen/utils/ErrorCode.h"
 
 namespace meen
@@ -429,17 +431,23 @@ std::error_code Intel8080::Load(const std::string&& str, bool checkUuid)
 std::expected<std::string, std::error_code> Intel8080::Save() const
 {
 	auto b64 = Utils::BinToTxt("base64", "none", uuid_.data(), uuid_.size());
-	
+
 	if (!b64)
 	{
 		return b64;
 	}
 
-	auto fmtStr = R"({"uuid":"base64://%s","registers":{"a":%d,"b":%d,"c":%d,"d":%d,"e":%d,"h":%d,"l":%d,"s":%d},"pc":%d,"sp":%d})";
-	auto count = snprintf(nullptr, 0, fmtStr, b64.value().c_str(), Value(a_), Value(b_), Value(c_), Value(d_), Value(e_), Value(h_), Value(l_), Value(status_), pc_, sp_);
-	std::string str(count + 1, '\0');
-	snprintf(str.data(), count + 1, fmtStr, b64.value().c_str(), Value(a_), Value(b_), Value(c_), Value(d_), Value(e_), Value(h_), Value(l_), Value(status_), pc_, sp_);
-	return str;
+	auto a = Value(a_);
+	auto b = Value(b_);
+	auto c = Value(c_);
+	auto d = Value(d_);
+	auto e = Value(e_);
+	auto h = Value(h_);
+	auto l = Value(l_);
+	auto s = Value(status_);
+
+	return std::vformat(R"({{"uuid":"base64://{}","registers":{{"a":{},"b":{},"c":{},"d":{},"e":{},"h":{},"l":{},"s":{}}},"pc":{},"sp":{}}})",
+						std::make_format_args(b64.value(), a, b, c, d, e, h, l, s, pc_, sp_));
 }
 #endif // ENABLE_MEEN_SAVE