+++ source/machine/Machine.cpp
@@ -24,11 +24,12 @@ SOFTWARE.
 #include <bit>
 #include <charconv>
 #include <cinttypes>
+#include <format>
 #include <numeric>
 #include <stdio.h>
-#ifdef ENABLE_MEEN_RP2040
+#ifdef PICO_BOARD
 #include <pico/multicore.h>
-#endif // ENABLE_MEEN_RP2040
+#endif // PICO_BOARD
 #ifdef ENABLE_NLOHMANN_JSON
 #include <nlohmann/json.hpp>
 #else
@@ -125,10 +126,10 @@ namespace meen
 		int64_t totalTicks = 0;
 		int64_t lastTicks = 0;
 		auto ticksPerIsr = 0ull;
-#ifndef ENABLE_MEEN_RP2040
+#ifndef PICO_BOARD
 		auto loadLaunchPolicy = m->opt_.LoadAsync() ? std::launch::async : std::launch::deferred;
 		std::future<std::string> onLoad;
-#endif // ENABLE_MEEN_RP2040
+#endif // PICO_BOARD
 		auto loadMachineState = [&](std::string&& str)
 		{
 			if (str.empty() == false)
@@ -173,9 +174,14 @@ namespace meen
 						fseek(fin, 0, SEEK_END);
 						std::string jsonStr(ftell(fin), '\0');
 						fseek(fin, 0, SEEK_SET);
-						fread(jsonStr.data(), 1, jsonStr.size(), fin);
-						fclose(fin);
+						
+						if (fread(jsonStr.data(), jsonStr.size(), 1, fin) != 1)
+						{
+							fclose(fin);
+							return m->HandleError(errc::incompatible_rom, std::source_location::current());
+						}
 
+						fclose(fin);
 						auto je = deserializeJson(json, jsonStr);
 
 						if (je)
@@ -584,7 +590,7 @@ namespace meen
 						{
 							scheme = memory["rom"]["scheme"].get<std::string_view>();
 						}
-				
+
 						if (memory["rom"].contains("directory"))
 						{
 							directory = memory["rom"]["directory"].get<std::string_view>();
@@ -785,7 +791,7 @@ namespace meen
 			return std::error_code{};
 		};
 
-#ifndef ENABLE_MEEN_RP2040
+#ifndef PICO_BOARD
 		auto checkHandler = [](std::future<std::string>& fut)
 		{
 			std::string str;
@@ -802,7 +808,7 @@ namespace meen
 
 			return str;
 		};
-#endif // ENABLE_MEEN_RP2040
+#endif // PICO_BOARD
 #ifdef ENABLE_MEEN_SAVE
 		auto saveLaunchPolicy = m->opt_.SaveAsync() ? std::launch::async : std::launch::deferred;
 		std::future<std::string> onSave;
@@ -833,18 +839,18 @@ namespace meen
 				{
 					// If a user defined callback is set and we are not processing a load or save request
 					if (m->onLoad_ != nullptr
-#ifndef ENABLE_MEEN_RP2040
+#ifndef PICO_BOARD
 						&& onLoad.valid() == false
-#endif // ENABLE_MEEN_RP2040
+#endif // PICO_BOARD
 #ifdef ENABLE_MEEN_SAVE
 						&& onSave.valid() == false
 #endif // ENABLE_MEEN_SAVE
 					)
 					{
-#ifndef ENABLE_MEEN_RP2040
+#ifndef PICO_BOARD
 						onLoad = std::async(loadLaunchPolicy, [m]
 						{
-#endif // ENABLE_MEEN_RP2040
+#endif // PICO_BOARD
 							int len = m->opt_.MaxLoadStateLength();
 							std::string str(len, '\0');
 
@@ -857,14 +863,14 @@ namespace meen
 							}
 
 							str.resize(len);
-#ifndef ENABLE_MEEN_RP2040
+#ifndef PICO_BOARD
 							return str;
 						});
 
 						loadMachineState(checkHandler(onLoad));
 #else
 						loadMachineState(std::move(str));
-#endif // ENABLE_MEEN_RP2040
+#endif // PICO_BOARD
 					}
 					break;
 				}
@@ -909,37 +915,12 @@ namespace meen
 
 									if (cpuStateTxt)
 									{
-										size_t dataSize = 0;
-										auto fmtStr = R"({"cpu":%s,"memory":{"uuid":"%s://%s","rom":{"bytes":"%s://md5://%s"},"ram":{"size":%d,"bytes":"%s://%s://%s"}}})";
+										auto ramSize = ram.size();
+										auto str = std::vformat(R"({{"cpu":{},"memory":{{"uuid":"{}://{}","rom":{{"bytes":"{}://md5://{}"}},"ram":{{"size":{},"bytes":"{}://{}://{}"}}}}}})",
+																std::make_format_args(cpuStateTxt.value(), m->opt_.Encoder(), memUuidTxt.value(), m->opt_.Encoder(), romMd5Txt.value(),
+																ramSize, m->opt_.Encoder(), m->opt_.Compressor(), ramTxt.value()));
 
-										// todo - replace snprintf with std::format
-										auto writeState = [&]()
-										{
-											std::string str;
-											char* data = nullptr;
-
-											if (dataSize > 0)
-											{
-												str.resize(dataSize);
-												data = str.data();
-											}
-
-											//cppcheck-suppress nullPointer
-											dataSize = snprintf(data, dataSize, fmtStr,
-												cpuStateTxt.value().c_str(),
-												m->opt_.Encoder().c_str(),
-												memUuidTxt.value().c_str(),
-												m->opt_.Encoder().c_str(),
-												romMd5Txt.value().c_str(),
-												ram.size(), m->opt_.Encoder().c_str(), m->opt_.Compressor().c_str(),
-												ramTxt.value().c_str()) + 1;
-
-											return str;
-										};
-
-										writeState();
-
-										onSave = std::async(saveLaunchPolicy, [m, state = writeState()]
+										onSave = std::async(saveLaunchPolicy, [m, state = std::move(str)]
 										{
 											int len = 255;
 											std::string uri(len + 1, '\0');
@@ -963,7 +944,7 @@ namespace meen
 													if (!ec)
 													{
 														std::ofstream fout(path / fileName, std::ios::trunc);
-														
+
 														if (!fout.good())
 														{
 															e = meen::errc::invalid_argument;
@@ -990,7 +971,7 @@ namespace meen
 													{
 														e = meen::errc::invalid_argument;
 													}
-													
+
 													if (e)
 													{
 														// We failed to write, give the completion handler a chance to write it correctly.
@@ -1057,13 +1038,13 @@ namespace meen
 				case ISR::Quit:
 				{
 					// Wait for any outstanding load/save requests to complete
-#ifndef ENABLE_MEEN_RP2040
+#ifndef PICO_BOARD
 					if (onLoad.valid() == true)
 					{
 						// we are quitting, wait for the onLoad handler to complete
 						loadMachineState(onLoad.get());
 					}
-#endif // ENABLE_MEEN_RP2040
+#endif // PICO_BOARD
 #ifdef ENABLE_MEEN_SAVE
 					if (onSave.valid() == true)
 					{
@@ -1097,9 +1078,9 @@ namespace meen
 							quit = m->onIdle_(m->ioController_.get());
 						}
 					}
-#ifndef ENABLE_MEEN_RP2040
+#ifndef PICO_BOARD
 					loadMachineState(checkHandler(onLoad));
-#endif // ENABLE_MEEN_RP2040
+#endif // PICO_BOARD
 #ifdef ENABLE_MEEN_SAVE
 					checkHandler(onSave);
 #endif // ENABLE_MEEN_SAVE
@@ -1198,7 +1179,7 @@ namespace meen
 		{
 			if(running_ == true)
 			{
-#ifdef ENABLE_MEEN_RP2040
+#ifdef PICO_BOARD
 				auto core1Ret = multicore_fifo_pop_blocking();
 
 				if(core1Ret != 0xFFFFFFFF)
@@ -1214,7 +1195,7 @@ namespace meen
 				{
 					return HandleError(errc::async, std::source_location::current());
 				}
-#endif // ENABLE_MEEN_RP2040
+#endif // PICO_BOARD
 				running_ = false;
 			}
 
@@ -1236,7 +1217,7 @@ namespace meen
 			}
 		};
 
-#ifdef ENABLE_MEEN_RP2040
+#ifdef PICO_BOARD
 		if(opt_.RunAsync() == true)
 		{
 			auto runMachineAsync = []
@@ -1290,7 +1271,7 @@ namespace meen
 				return std::unexpected(err);
 			}
 		}
-#endif // ENABLE_MEEN_RP2040
+#endif // PICO_BOARD
 		return runTime_;
 	}